Correct support of fonts for card name and subname
The UI was showing the ability to set the font, style and size for both the card name and sub name, but it wasn't storing the style for the card name, nor was it saving the font settings for the subname. Also, the font selected for the cardname was also being applied unto the subname.