Fix for #20 Control for how much space a line break takes
Added some properties to Text Areas and Scrolling Text Area to control the space between lines and when entering a manual line break.
Since the code by default uses a value of 0.6 for manual line breaks, I left it like that and instead modified the Heroes and Scheme templates to set that value to 0.2 instead (which is the value used for between lines).
This way anyone can tweak them if they like. XML properties are called spacebetweenlines and spacebetweenparagraphs.