New Card Types and clean up (Villain Bystander, Operations, Custom Scheme Twist)
Added 3 new card types, as provided by Tovarich Pizann:
Bystander Villain: A Bystander that enters the city as a Villain when rescued.
Custom Scheme Twist: Scheme twist with custom text. Can be used for different game effects.
Operation: This is a Scheme where the players have to do an extra thing in addition to defeating the Mastermind in order to win. In short, same as schemes but with a different name. Can be used to differentiate with normal scheme and use them as extra game setup.
Adjusted all bystanders, wounds and bindings cards to have the Card type properly aligned with their banner (all of them were displaying their card type text too low and outside of the banner from the image template).
Adjusted Scheme and Scheme twist cards to start displaying their card text a bit higher. By default, the firt line would be so low on the card it could actually be cut out.
Removed some underlay images from some card types that weren't used by them.