QoL: Consider merging some templates together #49
Implemented basetemplate that allows a template to refer to another template as a base for itself. This in turn allows for way simpler templates as you don't have to copy over a full template just for the sake of modifying a small part of it.
Converted several templates to use the new functionality.